requirements.txt
django==4.2
djangorestframework-simplejwt>=5.2.0
psycopg2==2.9.9
psycopg2-binary>=2.9.3
django-stubs==5.0.2
django-environ==0.11.2
python-decouple==3.8
django-cors-headers==4.4.0
djangorestframework==3.15.2
django-allauth>=0.51.0
django-rest-auth>=0.9.5
gunicorn>=20.1.0
# python-decouple>=3.6
pillow==10.4.0
django-storages==1.14.4
django-ckeditor-5==0.2.13
coreapi==2.3.3

